requirements.txt
streamlit
openai-whisper
transformers
torch
torchaudio
pydub
numpy
ffmpeg-python
ffmpeg-python
